code_runner/requirements/requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flower==0.9.5
humanize==3.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
